Updated week two node func
