+++ k8s-knative-classifier-course/trigger-func/Dockerfile
@@ -9,6 +9,7 @@ WORKDIR /usr/src/app
 # A wildcard is used to ensure both package.json AND package-lock.json are copied.
 # Copying this separately prevents re-running npm install on every code change.
 COPY package*.json ./
+COPY sglynnbot-key.json ./
 
 # Install production dependencies.
 RUN npm install --only=production
