+++ k8s-knative-classifier-course/trigger-func/package.json
@@ -10,6 +10,7 @@
   "license": "Apache-2.0",
   "dependencies": {
     "@google-cloud/pubsub": "^2.10.0",
+    "body-parser": "^1.18.3",
     "dotenv": "^8.2.0",
     "express": "^4.16.4",
     "firebase-admin": "^9.5.0",
